test_requirements.txt
django==2.2.0
snowland-djangohelper>=0.1.1
django-shortuuidfield>=0.1.3
djangorestframework==3.9.1
idna==2.7
mysqlclient==1.4.0
Pillow==5.1.0
PyMySQL==0.8.1
pytz==2018.4
requests==2.20.0
shortuuid==0.5.0
six==1.11.0
urllib3==1.23
django_debug_toolbar
django-silk
snowland-smx
djdt_flamegraph
qiniu>=7.3.1
astartool>0.0.7
